Backend/requirements.txt
Flask
vertexai
google-cloud-aiplatform
google-cloud